Cleanup and minor refactoring
