+++ src/Validator.php
@@ -80,8 +80,11 @@ final class Validator implements ValidatorInterface
         foreach ($rules ?? [] as $attribute => $attributeRules) {
             $result = new Result();
 
-            $tempRule = is_iterable($attributeRules) ? $attributeRules : [$attributeRules];
-            $attributeRules = $this->normalizeRules($tempRule);
+            if (!is_iterable($attributeRules)) {
+                $attributeRules = [$attributeRules];
+            }
+
+            $attributeRules = $this->normalizeRules($attributeRules);
 
             if (is_int($attribute)) {
                 $validatedData = $data->getData();
