+++ tests/Stub/ObjectWithCallbackMethod/ObjectWithNonExistingCallbackMethod.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Validator\Tests\Stub;
+namespace Yiisoft\Validator\Tests\Stub\ObjectWithCallbackMethod;
 
 use Yiisoft\Validator\Rule\Callback;
