+++ tests/Stub/PiHandler.php
@@ -10,15 +10,12 @@ use Yiisoft\Validator\ValidationContext;
 
 final class PiHandler implements RuleHandlerInterface
 {
-    /**
-     * @inheritDoc
-     */
-    public function validate(mixed $value, object $rule, ?ValidationContext $context = null): Result
+    public function validate(mixed $value, object $rule, ValidationContext $context): Result
     {
         $result = new Result();
 
         if (!(\abs($value - M_PI) < PHP_FLOAT_EPSILON)) {
-            $result->addError('Value must be Pi.', ['value' => $value]);
+            $result->addError('Value must be Pi.');
         }
 
         return $result;
